Make docker_explorer a proper python module and move the main script into 'tools'
