+++ docker_explorer/__init__.py
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """docker-explorer tool."""
-__version__ = '20180612'
+__version__ = '20190306'
